+++ src/settings/_index.njk
@@ -21,20 +21,6 @@
   </h2>
 
   <h3>
-    Import Data
-  </h3>
-  <p>
-    If you've exported data from another device or browser, you can import it
-    here. It should have a name similar to
-    "cash-cache.<span data-current-date></span>.json".
-  </p>
-  <div>
-    <label for="importFile" class="button">Choose a File</label>
-    <input id="importFile" type="file" name="import-file" class="visually-hidden" data-import-data>
-  </div>
-  <div data-import-progress aria-live="polite"></div>
-
-  <h3>
     Delete Data
   </h3>
   <p>
