+++ src/settings/import-data.njk
@@ -15,8 +15,19 @@
       Go Back
     </a>
   </div>
+
+  <p>
+    If you've exported data from another device or browser, you can import it
+    here. It should have a name similar to
+    "cash-cache.<span data-current-date></span>.json".
+  </p>
+  <div>
+    <label for="importFile" class="button">Choose a File</label>
+    <input id="importFile" type="file" name="import-file" class="visually-hidden" data-import-data>
+  </div>
+  <div data-import-progress aria-live="polite"></div>
 {% endblock %}
 
 {% block scripts %}
-  <script src="/scripts/settings/import-data.js"></script>
+  <script src="/scripts/settings/import-data.mjs" type="module"></script>
 {% endblock %}
